bibliotekapi/requirements.txt
dependency-injector==4.42.0
fastapi==0.115.4
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
SQLAlchemy==2.0.45
uvicorn==0.32.0
email-validator==2.3.0
asyncpg==0.29.0
python-multipart
